A few bugfixes in the ReWeight engine. ReWeight engine now uses reduced event format by default. Changes how LFGKfScale is interpreted to be more physical. Added parameter OutputEvtHistogram which triggers output of NEUT-like run summary histograms for uniform beam profiles. Some const-ifying of the core classes.
  - AzeemsAnalysis/TheFittening.cxx:
      - Tweaks to phase space cuts.

  - src/RooTrackerEvent.cc:
    - Much more robust NuWro event -> NEUT code implementation, ported here from
      elsewhere in the source.
    - Now properly looks at the final states to determin more granular NEUT
      codes.
    - Re-indent.

  - src/nuwro2rootracker.cc:
    - Quientens progress polling rate by a factor of 100.
    - Re-indent.

  - src/params.xml;
    - Fixes the default and description of nucleus_LFGkfScale

  - src/EnergyProfile.h:
    - Re-indent.

  - src/beam_uniform.h:
    - size_t for loop counters of std::vectors.
    - Re-indent.

  - src/kaskada7.cc:
    - Re-indent.

  - src/nucleus.cc:
    - Adjusts how the LFGKfScale parameter is interpreted.
      - It is now a dimensionless scale on the nominal kF as calculated from
        the LFG.
      - Effectively moves the 'Fermi surface' of the momentum distribution.
      - Currently very 'for testing only'.

  - src/nuwro.cc:
    - Adds a reaction to the new parameter OutputEvtHistogram.
      - If enabled histograms of the flux shape, the total flux averaged cross
        section and the total event rate in ENu are output with the event file.
      - This is to help unify code bases with how programs dealing with NEUT
        events calculate their event-by-event weight.
      - No change to the event by event weight code, just extra histograms in the
        output file.
    - Re-indent

  - src/params.h:
    - removes some unused local variables that were causing warnings.

  - src/particle.h:
    - Adds a const version of p4.

  - src/target_mixer.cc:
    - Re-indent.

  - src/reweight/NuwroReWeight.{h,cc}:
    - Adds splash logo... (only for gcc).
    - Re-indent.
    - Adds a hard check for attempts to set non-normal (inf/nan) dial values.
    - Adds calc interface which uses SimpleReWeightEvents.
    - Adds method to set reweight engine nominals from a params instance.

  - src/reweight/NuwroReWeightSimple.h:
    - Adds 2p2hNorm dial to SimpleReWeightParams struct.

  - src/reweight/NuwroReWeightSimpleEvent.h:
    - Adds a mutable cacheweight to the SRWEvent format, if not set to a default
      this will be assumed to be the events generated 'old weight', reduces the
      need to pass such weight caches into the engine manually.
    - Adds pretty printing of SRWEvents.

  - src/reweight/NuwroReWeight_BBBA07.{h,cc}:
    - Re-indent.

  - src/reweight/NuwroReWeight_FlagNorm.h:
    - Re-indent.

  - src/reweight/NuwroReWeight_QEL.cc:
    - Returns MaCCQE value definition to convention where errors are dimless
      fractions.
      -  fCurr_MaCCQE = fDef_MaCCQE + fError_MaCCQE * fTwkDial_MaCCQE;
      +  fCurr_MaCCQE = fDef_MaCCQE * (1.0 + fError_MaCCQE * fTwkDial_MaCCQE);
    - The SRWEvent interface is now the default and the event interface is
      implemented through a temporary SRWEvent.
      - I don't this is any slower as all the SRWEvent constructor does is get
        the information we need for reweighting the given event type.
    - Changes check to QE events to use SRWEvent format.

  - src/reweight/NuwroReWeight_SPP.cc:
    - Fixes a units bug in the default value of MaRES.
    - Implements the SetDefault method for setting up engine defaults from a
      params instance.
    - Returns dial value definition to convention where errors are dimless
      fractions.
    - Changes check to SPP events to use SRWEvent format.
    - Implements SRWEvent cache weight usage.

  - src/reweight/NuwroSystUncertainty.cc:
    - Fixes a units bug in the default errors of MaRES.

  - src/reweight/NuwroWghtEngineI.h:
    - Re-indent.
    - Adds CalcWeight(SRW::SRWEvent const &) method.
      - Comes with default implementation for engines that do not implement it
        yet.
      - This does currently just return a weight of one...

  - src/vali/SAF_Analysis.cc:
    - Also outputs the correctly averaged event rate per event for simpler
      loading of subsets of events.

  - src/vali/SimpleAnalysisFormat.h:
    - Re-indent.

  - src/vali/reweight/apps/SimpleReweighter.cc:
    - Now uses SRWEvent format.
    - Re-indent.

The are other specific or uninteresting vali script changes.

+++ src/event1.h
@@ -217,7 +217,7 @@ int event::f() const { return post.size(); }
 /// number of particles with given pdg after primary vertex
 int event::nof(int pdg)const  {
   int c = 0;
-  for (int i = 0; i < out.size(); i++)
+  for (size_t i = 0; i < out.size(); i++)
     if (out[i].pdg == pdg) c++;
   return c;
 }
@@ -225,7 +225,7 @@ int event::nof(int pdg)const  {
 /// number of particles with given pdg after primary vertex
 int event::nof(int pdg1, int pdg2) const {
   int c = 0;
-  for (int i = 0; i < out.size(); i++)
+  for (size_t i = 0; i < out.size(); i++)
     if (out[i].pdg == pdg1 || out[i].pdg == pdg2) c++;
   return c;
 }
@@ -233,7 +233,7 @@ int event::nof(int pdg1, int pdg2) const {
 /// number of particles with given pdg after primary vertex
 int event::nof(int pdg1, int pdg2, int pdg3) const {
   int c = 0;
-  for (int i = 0; i < out.size(); i++)
+  for (size_t i = 0; i < out.size(); i++)
     if (out[i].pdg == pdg1 || out[i].pdg == pdg2 || out[i].pdg == pdg3) c++;
   return c;
 }
@@ -241,7 +241,7 @@ int event::nof(int pdg1, int pdg2, int pdg3) const {
 /// number of particles with given pdg leaving nucleus
 int event::fof(int pdg) const {
   int c = 0;
-  for (int i = 0; i < post.size(); i++)
+  for (size_t i = 0; i < post.size(); i++)
     if (post[i].pdg == pdg) c++;
   return c;
 }
@@ -249,7 +249,7 @@ int event::fof(int pdg) const {
 /// number of particles with given pdg leaving nucleus
 int event::fof(int pdg1, int pdg2) const {
   int c = 0;
-  for (int i = 0; i < post.size(); i++)
+  for (size_t i = 0; i < post.size(); i++)
     if (post[i].pdg == pdg1 || post[i].pdg == pdg2) c++;
   return c;
 }
@@ -257,7 +257,7 @@ int event::fof(int pdg1, int pdg2) const {
 /// number of particles with given pdg leaving nucleus
 int event::fof(int pdg1, int pdg2, int pdg3) const {
   int c = 0;
-  for (int i = 0; i < post.size(); i++)
+  for (size_t i = 0; i < post.size(); i++)
     if (post[i].pdg == pdg1 || post[i].pdg == pdg2 || post[i].pdg == pdg3) c++;
   return c;
 }
@@ -279,14 +279,14 @@ int event::charge(int r) const {
       c = pr;
       break;
   }
-  for (int i = 0; i < set->size(); i++) c += (*set)[i].charge();
+  for (size_t i = 0; i < set->size(); i++) c += (*set)[i].charge();
   return c;
 }
 
 /// invaraint mass
 double event::W() const {
   vect h = out[1];
-  for (int a = 2; a < out.size(); a++) h = h + out[a];
+  for (size_t a = 2; a < out.size(); a++) h = h + out[a];
   return sqrt(h * h);
 }
 
@@ -294,12 +294,12 @@ double event::W() const {
 double event::przod() const {
   int licz = 0;
   vect h = out[1];
-  for (int a = 2; a < out.size(); a++) h = h + out[a];
+  for (size_t a = 2; a < out.size(); a++) h = h + out[a];
 
   vect tran = in[0] - out[0];
   vec ptr = vec(tran.x, tran.y, tran.z);
 
-  for (int b = 1; b < out.size(); b++) {
+  for (size_t b = 1; b < out.size(); b++) {
     if ((out[b].pdg != 111) && (out[b].pdg != 22) && (out[b].pdg != 130) &&
         (out[b].pdg != 2112) && (out[b].pdg != -2112))
 
@@ -319,12 +319,12 @@ double event::przod() const {
 double event::tyl() const {
   int licz = 0;
   vect h = out[1];
-  for (int a = 2; a < out.size(); a++) h = h + out[a];
+  for (size_t a = 2; a < out.size(); a++) h = h + out[a];
 
   vect tran = in[0] - out[0];
   vec ptr = vec(tran.x, tran.y, tran.z);
 
-  for (int b = 1; b < out.size(); b++) {
+  for (size_t b = 1; b < out.size(); b++) {
     if ((out[b].pdg != 111) && (out[b].pdg != 22) && (out[b].pdg != 130) &&
         (out[b].pdg != 2112) && (out[b].pdg != -2112))
 
@@ -366,7 +366,7 @@ int event::number_of_escape() const { return nod[10]; }
 
 int event::number_of_interactions() const {
   int noi = 0;
-  for (int i = 0; i < 11; i++) {
+  for (size_t i = 0; i < 11; i++) {
     noi = noi + nod[i];
   }
 
@@ -378,11 +378,11 @@ int event::number_of_particles(int pdg, bool fsi)const  {
   int number = 0;
 
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if (post[k].pdg == pdg) number++;
     }
   } else if (!fsi) {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if (out[k].pdg == pdg) number++;
     }
   }
@@ -393,7 +393,7 @@ int event::number_of_particles(int pdg, bool fsi)const  {
 /// total kinetic energy of nucleons that left the nucleus
 double event::nuc_kin_en() const {
   double sum = 0;
-  for (int k = 0; k < post.size(); k++) {
+  for (size_t k = 0; k < post.size(); k++) {
     if (post[k].pdg == 2212 || post[k].pdg == 2112) sum += post[k].Ek();
   }
   return sum;
@@ -405,11 +405,11 @@ int event::num_part_thr(int pdg, bool fsi, double threshold) const {
   int number = 0;
 
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if ((post[k].pdg == pdg) && (post[k].momentum() > threshold)) number++;
     }
   } else if (!fsi) {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if ((out[k].pdg == pdg) && (out[k].momentum() > threshold)) number++;
     }
   }
@@ -422,7 +422,7 @@ double event::proton_cosine(bool fsi, double thr)const  {
   int numer[2];
   int ile = 0;
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if (post[k].pdg == 2212 && post[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -434,7 +434,7 @@ double event::proton_cosine(bool fsi, double thr)const  {
     // )/post[numer[0]].momentum()/post[numer[1]].momentum();
 
   } else {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if (out[k].pdg == 2212 && out[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -450,9 +450,9 @@ double event::proton_cosine(bool fsi, double thr)const  {
 /// calculates momentum of proton which did not suffer from fsi (the first one,
 /// sometimes there are two,see below)
 double event::proton_transp_mom() const {
-  for (int k = 0; k < out.size(); k++)
+  for (size_t k = 0; k < out.size(); k++)
     if (out[k].pdg == 2212) {
-      for (int l = 0; l < post.size(); l++) {
+      for (size_t l = 0; l < post.size(); l++) {
         if (post[l].pdg == 2212) {
           //				double kos = ( out[k].x*post[l].x +
           //out[k].y*post[l].y + out[k].z*post[l].z
@@ -469,9 +469,9 @@ double event::proton_transp_mom() const {
 /// if it happens -- very rarely, but stil...)
 double event::proton_transp_mom2() const {
   int counter = 0;
-  for (int k = 0; k < out.size(); k++)
+  for (size_t k = 0; k < out.size(); k++)
     if (out[k].pdg == 2212) {
-      for (int l = 0; l < post.size(); l++) {
+      for (size_t l = 0; l < post.size(); l++) {
         if (post[l].pdg == 2212) {
           double kos = (out[k].x * post[l].x + out[k].y * post[l].y +
                         out[k].z * post[l].z) /
@@ -488,9 +488,9 @@ double event::proton_transp_mom2() const {
 /// calculates number of protons which did not suffer from fsi
 int event::proton_transp() const {
   int ile = 0;
-  for (int k = 0; k < out.size(); k++)
+  for (size_t k = 0; k < out.size(); k++)
     if (out[k].pdg == 2212) {
-      for (int l = 0; l < post.size(); l++) {
+      for (size_t l = 0; l < post.size(); l++) {
         if (post[l].pdg == 2212) {
           //					double kos = ( out[k].x*post[l].x +
           //out[k].y*post[l].y + out[k].z*post[l].z
@@ -507,7 +507,7 @@ int event::proton_pair_number1(bool fsi, double thr) const {
   int numer[2];
   int ile = 0;
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if (post[k].pdg == 2212 && post[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -515,7 +515,7 @@ int event::proton_pair_number1(bool fsi, double thr) const {
     }
     return numer[0];
   } else {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if (out[k].pdg == 2212 && out[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -529,7 +529,7 @@ int event::proton_pair_number2(bool fsi, double thr) const {
   int numer[2];
   int ile = 0;
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if (post[k].pdg == 2212 && post[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -537,7 +537,7 @@ int event::proton_pair_number2(bool fsi, double thr) const {
     }
     return numer[1];
   } else {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if (out[k].pdg == 2212 && out[k].momentum() > thr) {
         numer[ile] = k;
         ile++;
@@ -553,12 +553,12 @@ double event::part_max_mom(int pdg, bool fsi) const {
   double mom = 0.0;
 
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if ((post[k].pdg == pdg) && (post[k].momentum() > mom))
         mom = post[k].momentum();
     }
   } else if (!fsi) {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if ((out[k].pdg == pdg) && (out[k].momentum() > mom))
         mom = out[k].momentum();
       // cout<<mom<<endl;
@@ -575,7 +575,7 @@ double event::part_sec_mom(int pdg, bool fsi) const {
   double memory;
 
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if (post[k].pdg == pdg) {
         double ped = post[k].momentum();
         if (ped > mom[1]) {
@@ -589,7 +589,7 @@ double event::part_sec_mom(int pdg, bool fsi) const {
       }
     }
   } else if (!fsi) {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if (out[k].pdg == pdg) {
         double ped = out[k].momentum();
         // cout<<ped<<endl;
@@ -613,7 +613,7 @@ double event::vert_act(double pion_threshold, bool fsi,
   double veract = 0.0;
 
   if (fsi) {
-    for (int k = 0; k < post.size(); k++) {
+    for (size_t k = 0; k < post.size(); k++) {
       if ((post[k].pdg == 211) && (post[k].momentum() < pion_threshold))
         veract += post[k].t - post[k].mass();
 
@@ -628,7 +628,7 @@ double event::vert_act(double pion_threshold, bool fsi,
     }
     return veract;
   } else {
-    for (int k = 0; k < out.size(); k++) {
+    for (size_t k = 0; k < out.size(); k++) {
       if ((out[k].pdg == 211) && (out[k].momentum() < pion_threshold))
         veract += out[k].t - out[k].mass();
 
@@ -657,7 +657,7 @@ double event::Erec(double Bin) const {
 /// stop program if event weight or momentum of any particle is NaN (not a
 /// number)
 void event::check() const {
-  for (int i = 0; i < out.size(); i++) {
+  for (size_t i = 0; i < out.size(); i++) {
     particle p = out[i];
     if (!(p.x == p.x && p.y == p.y && p.z == p.z and p.t == p.t)) {
       cerr << p << endl;
